Add includes feature
This adds support for including other Cherri files in a Cherri file.
This essentially just pastes the included file into your files contents
just before parsing. The path must exist and be a .cherri file.

To add this feature this refactors how we check files so that we don't
repeat the same file checks.