Refactor input variable to alias function for actions
This refactors what we use for "WFInput" to an alias function, we were
giving an empty string which to this function meant to get the UUID of
the input from the uuid map of variable UUIDs, but this looked really
silly so this was refactored to an alias function that will use the uuid
for the variable name if it is found.

That's another thing, this will now check to make sure we have a UUID
for that variable and if not just use an empty string, better to have
the wrong answer than to crash!