Fully implement optional arguments
This completes the implementation of optional arguments.

This adds an optional property to argument definitions.

This takes defaultValue and optional into account when writing the
actual key value pair for the parameter for the action.

I wonder if defaultValue is even necessary? The Shortcuts app doesn't
write values not specified. We may just need to refactor actions and
replace defaultValue with optional.