Organize actions
This renames the actions file to actions_standard.

This also renames the function to standardAction() and makes the actions
map when we start parsing as it's likely we'll need it.

This is done before adding all of the standard actions for organization
purposes and to signify intent, other types of actions should go in
newer actions_EXAMPLE.go files