Refine collecting until
This creates a function decended from the collectUntil(ch) function.
collectUntilExpect throws an expected character error and returns what
we collected if we are to advance past `maxAdvances` and not find `ch`
beforehand.

This is done to stabilize Cherri more, and commits will follow using it
in more places. This will prevent Cherri from thinking something goes on
forever and never actually stopping until we reach the end, which could
take a long time.

This adds documentation comments as well.