Fix and improve debug output
Since we started freeing up memory between steps in the compiler, the
previous debug output has been rendered useless.

This breaks up the debug output so that when we finish a step we print
debug info relating to that step before we unset the data we stored that
we no longer need now that the step is complete.

These functions are combined into a panicDebug() function when we have a
parser error or need to exit for some reason so we can see any data the
compiler has stored so far when we error out.