bugfix #22: Improve add to variable syntax
This fixes instances where += syntax is used on a number or string.

In the case of an integer, +=, -=, *= and /= may be used to modify the
variable with a literal or variable using the Calculate action.

In the case of a string, we reassign the variable using a text action to
append the variable or literal to the end of the original variable.

Uninitialized values will be added to using the type of variable that is
being added to them.

Arrays will continue to create a 'Add to Variable' action to add
whatever value to that variable.