Unify checking bundle ids and regex identifiers
This unifys the logic with checking for a replacement bundle ID. This
also adds a regex check against bundle IDs that don't match the
replacements map, throwing an error if it doesn't match a valid bundle
identifier.

Maybe we can change this to a warning in the future, but for now this
seems like something we should throw an error for so the user knows the
compiled Shortcut will not work.