Refactor mac definition to use definitions map
This is mainly to not have a default value for a definition by having it
stored in a global. I ran into this where the compiler complains I can't
use an action because it is a macOS only action. This should only happen
if I use `#define mac false`, I should have to specifically say I want
it to throw an error when I use a mac action in my non-mac shortcut.