Refactor the way actions are defined
This is a major refactor of how actions are defined.

Previously, we were defining them in multiple long functions and this
was essentially to make sure that we only defined the actions map when
needed.

However, this makes for very long functions which is bad practice. This
is also not needed as most of the time actions will be used in a typical
Cherri file.