Implement HubSign as fallback for macOS signing
The folks at RoutineHub have been gracious enough to allow Cherri to use
their HubSign service to sign Shortcuts in the compiler.

We were doing this previously as a necessity in the playground, but now
the compiler will also fallback to using this service if macOS fails to
sign the Shortcut.

Obviously, this comes with a performance hit to compilation but is worth
it in the end.