Refactor variable reference handling
This refactors how we handle variable values since we weren't passing on
the "get" and "type" modifiers every time a variable value was used, we
mostly just passed the name of the variable around as a string.

I feel a bit nervous about this big of a variable value refactor but the
tests pass and it seems to look to be working when analyzing them.