Move first comment to top if using custom actions
If you are using a custom action, meaning we include the custom actions
header actions, when we are collecting actions, the first explicit
comment action (e.g. `comment("Hello")`) will be popped to the top of
the action tokens instead of the bottom.

This will only work for the first explicit comment action.

This is done to allow for instructional or contact comments that
developers would like to have at the top of their Shortcut even if they
are using custom actions, or at the very least the standard library.