Improve making custom action call dictionary
Wow. We currently make a JSON string and then make that into an
interface, rather than making a map which will also be more explicit.

I think we might have done this before due to previous plist generation
but I'm not sure why that would apply to this.

We now entirely make a map and default to the given value.