Move action definition parsing to preparsing
This is twofold, one to allow for actions to be defined whereever you
want without needing to worry about using it before you define it.

This also will allow for us to run this selectively, as if we define
actions outside of the compiler, we will probably need to be able to do
this arbitrarily.