Reset tokens when resetting parsing
As far as I was aware, this should be fine. The point of this function
anyway is to maintain the changes made to the contents, but completely
reset parsing as though we haven't been collecting anything, so if
anything this might be a good idea. This is primarily to make sure to
prevent comments from pre-parsing.